+++ src/main/java/com/galvanize/tmo/paspringstarter/BooksRepository.java
@@ -3,6 +3,6 @@ package com.galvanize.tmo.paspringstarter;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface BooksRepository extends JpaRepository<Books, int> {
+interface BooksRepository extends JpaRepository<Books, Long> {
 
 }
